Write temporary covers to `input_path`
Fixes #103. Right now if `input_path` is a directory rather than a file (say, `/input` as is recommended in the Docker setup), it will attempt to download the cover to `/input_cover.jpg` which will fail as described in #103. This fixes it so it will write to `/input/cover.jpg` (and doesn't change the current behaviour when `input_path` is a file)